requirements.txt
Django==3.2.7
djangorestframework==3.14.0

asgiref==3.8.1
dj-database-url==2.1.0
Django==3.2.7
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
psycopg2-binary==2.9.9
PyJWT==1.7.1
pytz==2024.1
setuptools==69.2.0
sqlparse==0.4.4
typing_extensions==4.10.0
